Latitude und Longitude auf Double umgestellt und Tests entsprechend angepasst

+++ src/main/scala/de.fhwedel.antscout/openstreetmap/GeographicCoordinate.scala
@@ -8,6 +8,6 @@ package openstreetmap
  * Time: 11:40
  */
 
-class GeographicCoordinate(val latitude: Float, val longitude: Float) {
+class GeographicCoordinate(val latitude: Double, val longitude: Double) {
     override def toString = "%f, %f" format (latitude, longitude)
 }
